refactor(iteration-loop): extract phases under 600-LOC ceiling (audit rec #4)
Audit recommendation #4 (iteration-loop.js part): the driver had 513
LOC. The audit suggested extracting 50+ line stages into named
functions. Done — iteration-loop.js now 311 LOC, with 5 phase
implementations sitting in their own files under ./iteration-phases/,
mirroring the established pre-loop-phases/ pattern.

## Layout

  iteration-loop.js                                  311 LOC (was 513)
    runSingleIteration                                   the iteration body
    runIterationLoop                                     the while loop
    [re-exports of the moved phases for back-compat]

  iteration-phases/coder-and-refactorer.js            48 LOC
    runCoderAndRefactorerStages

  iteration-phases/guards.js                          84 LOC
    runGuardStages (output + perf guards)

  iteration-phases/quality-gates.js                   76 LOC
    runQualityGateStages (TDD + Sonar)

  iteration-phases/reviewer-gate.js                   60 LOC
    runReviewerGateStage

  iteration-phases/handle-approved.js                 29 LOC
    handleApprovedReview

All phase files under the 150-LOC target; driver well under 350.

## Architecture tests repointed (no invariants relaxed)

Two architecture tests grep'd specific files for hardcoded patterns
that are still load-bearing — just at new addresses:

- tests/architecture/sonar-intrinsic.test.js
  Was: assert `iteration-loop.js` doesn't contain a forbidden Sonar
  toggle gate.
  Now: same assertion against `iteration-phases/quality-gates.js`.

- tests/architecture/stage-registry.test.js
  Was: assert iteration-loop.js calls `runStage(stageRegistry.get("coder"|"reviewer"))`.
  Now: same assertion against the two phase files where those calls
  now live.

The invariants are unchanged; the test files just track the new home.

Verified:
- npx vitest run --reporter=dot → 4192/4192 passing
- npx eslint src/               → 0 errors
